Properly check if PEP is supported (fdo #28230)
According to XEP-0163 PEP support is supposed to be check by discoing our own
bare JID. We got this wrong as we used to check the server features instead.
Recent ejabberd seems to not advertise PEP support in the server features any
more so we need to check it properly.