CapsCache: store entire disco reply in caps cache.
Previously we only stored the list of feature URIs included in the
disco reply for a particular hash. This isn't enough: we also make
decisions based on the <identity/> elements in the reply. For instance,
if we try to start an ad-hoc MUC, if one or more contact is using the
GMail client we use Google's PMUC server rather than the user's default
conference server. This information wasn't stored in the caps cache, so
was lost between sessions.

So, let's just store the entire <query/> node in the database.

This entailed a database schema migration. It's not much of a migration,
as such: if the database uses the old schema, it's just deleted.

In the process of arranging for obsolete databases to be deleted, I also
fixed a bunch of cases where broken databases wouldn't be removed. The
previous code tried to delete the database file and try again if
sqlite3_open() failed. This didn't catch:

 • corrupt databases, which only trigger failures when you run a query;
 • read-only databases, which only trigger failures when you run a query
   which modifies the database.

So this patch rearranges matters to try writing to the database before
declaring it okay.