A new world order! Restrung privacy list management.
When signing on, first check if there is a valid invisible list
before clobbering, this saves from push notification to other connected
resources. If the current 'invisible' list is not satisfactory, create an
'invisible-gabble' list that hopefully we could use.

This change also does not depend on privacy lists being a disco advertised
feature. All released versions of Prosody, and supposedly some version of
ejabberd, don't actually advertise their privacy list support. If we get an
error (presumably because of lack of support), the invisibility-via-xep-0126
is dropped in favor of xep-0018 or nothing.