muc-factory: don't associate a request with tubes if a tube was requested
Previously, the request token was associated with the tubes channel,
regardless of whether the request was for a tube or tubes
channel. This meant that if the MUC channel wasn't ready when a tube
channel was requested then the tubes channel will be given back to the
caller of CreateChannel. The MUC channel will be ready if a text MUC
has already been opened, where this bug does not appear.

Fixes: fd.o#25533

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>