Delete non-functional invisibility fallbacks test.
This test tested that if a server advertises (say) XEP-0186 support, but
sending an <invisible/> command fails, Gabble would fall back to trying
invisibility lists.

But, we removed that fallback, on the basis that if the server
advertises support for a feature, it should work. And
invisible_xep_0126.py should test any fallback from that to Þe Olde
<presence type='invisible'/>. Finally, this test didn't actually work
since InvisibleXmlStream was deleted. So, let's just bin it.