gabble_simple_async_countdown_new, _inc, _dec: new utility code
This encapsulates a common pattern for "plural" asynchronous operations
that are implemented as a set of parallel protocol operations: finish
only when all sub-operations have finished, with success if and
only if all sub-operations succeeded.

If this functionality proves to be useful, it should be added to
telepathy-glib, or even to GLib.