Make tests fail more gracefully when connections leak
Previously, if any test leaked its connection, it would fail (with a
reasonably helpful assertion). But the next test would try to re-make
the same connection, which would throw a DBusException (namely
NotAvailable). The exception wasn't handled, so it bubbled back up to
the mainloop, which would log it to stderr and then... do nothing. The
test would never finish, and it's a PITA to kill.

With this patch, we print a shorter representation of the NotAvailable
error and then exit immediately. This allows the rest of the test run to
finish. Since all the tests after the leaky one will probably fail,
let's add a note to the aforementioned helpful assertion to that effect
so the developer isn't too startled.