Port GabbleBaseCallChannel to TpBaseChannel
Unfortunately this makes jingle/test-wait-for-caps-incomplete.py reveal
a bug: if we're waiting for a contacts caps, but never receive them or
give up before we disconnect, async_init() on the channel never
finishes, so the channel is leaked. Since TpBaseChannel holds a ref to
its connection, this bug means the connection, too, is leaked; thus,
everyone dies.