Dramatically reduce debug spam from capabilities
With this patch, we only dump the full list of capabilities when we
decide to trust a caps uri; otherwise, we dump the diff being applied
to contacts. This still preserves all the information we used to show,
but means we don't dump the full set of caps 6 times whenever any
contact publishes a presence with a (possibly unchanged!) caps node.

Also, we no longer emit capabilities updates when nothing has changed.