Implemnent fallback-servers support
Implements support for the account parameter fallback-servers. This
parameter is an 'as' that contains a list of servers to try in the
form host[:port][,oldssl]. This is used for protocols such as Google
Talk where multiple ports and TLS authentications is being offered to
help firewall traversal. An example of fallback-servers array for
Google Talk would be:
    ["talkx.l.google.com",
     "talkx.l.google.com:443,oldssl",
     "talkx.l.google.com:80"]

Reviewed-by: Simon McVittie <simon.mcvittie@collabora.co.uk>