Always emit PresencesChanged when we change our own presence
For better or worse, if we're on a connection that supports privacy
lists, whenever we change our own status Gabble activates the
corresponding privacy list (which may be "no privacy list"), even if
this is the same one that's currently active.

Unfortunately, this code path never emitted PresencesChanged, leading to
<https://bugs.freedesktop.org/show_bug.cgi?id=30117>. This commit should
fix that, though it's not enough to make the test pass: there's a
reference leak somewhere.