TLSManager: don't crash if connection dies before verification
I received a report of a segfault with the following (trimmed) backtrace:

0  tp_base_channel_constructed () at base-channel.c:604
1  gabble_server_tls_channel_constructed (object=0x1619f8) at server-tls-channel.c:181
[ g_object_new crap ]
5  gabble_server_tls_manager_verify_async ([...]) at server-tls-manager.c:218
6  wocky_tls_handler_verify_async ([...]) at wocky-tls-handler.c:261
7  session_handshake_cb () at wocky-tls-connector.c:302

Line 604 of base-channel.c dereferences the channel's connection for the
first time. This is caused by the connection being NULL, which is
possible because ServerTLSManager's connection_status_changed_cb()
clears it when the connection moves to state DISCONNECTED (to break a
reference cycle). So, the crash occurs if you Disconnect() the
connection before Wocky gets as far as checking the certificate.

So, here we just fail the _verify_async() method call immediately if the
priv->connection is NULL.

Reviewed-by: Cosimo Cecchi <cosimo.cecchi@collabora.co.uk>