Stop listening to GabbleSlacker on stream error
Previously, GabbleConnection stopped listening to signals from MCE if it
was instructed to disconnect cleanly by the user, but not if the stream
ended abnormally. This patch amends its shut_down implementation to
rectify this. I believe this also fixes a leak of the WockyPing object
in the same situation; this would not lead to any crashes, but would
mean the WockyPorter leaks too.

Fixes: <https://bugs.freedesktop.org/show_bug.cgi?id=30207>