Make call contents' lifecycles more explicit.
When I reused some of GabbleBaseCallChannel, I found that it leaked
contents. This turned out to be because, in Gabble, the Channel's
reference to the content is released by the content itself!

So I made the ref/unref pairs more explicit. Now the content takes a
ref to itself while waiting for all its codec offers to die, and the
call channel explicitly releases its own refs.