Automate release process
The idea is that you smoke-test the release, then just run

  % make maintainer-make-release

which generates the tarball, the signature, the signed tag, and the
release mail, uploads the release to telepathy.freedesktop.org, and then
reminds you of the remaining steps (nano version bump, pushing to
upstream, and sending the mail).

If you want to do all the local bits, but not send anything, just run:

  % make maintainer-prepare-release

then do whatever you want to do with the tarball etc. If you think it's
ready, run:

  % make maintainer-upload-release