Split service-test from one queue into multiple named queues
A big source of race conditions in the test is because we're multiplexing
events from multiple event sources into one stream, depending on external
conditions the ordering of individual events can change. The gabble test go a
long way to try and cope with this, but it's nearly impossible to do so in all
cases.

This patch replaces the one big event queue by a queue per source and makes
the expect function only pop events from queues it's waiting on instead of all
of them. This makes the ordering of events as seen by the test completely
deterministic as it's now purely ordered by the ordering of the actual sources
not by the order in which we read from the sources.

For backwards compatibility the queue name is taken from the prefix of the
event currently, so the test don't need any changes. For more advanced test in
the future the queue names could be settable (e.g. different queues for
different dbus connections).

Some tests needed small updates because they got confused by earlier events
occuring on the queue they were interested in (mostly events that in the old
system would be dropped by execting events from unrelated sources).