client types: don't keep firing ClientTypesUpdated before we actually know anything
Previously, GabbleDiscoCache::presences-updated was emitted before the
disco request had been made to find out the actual client types of a
resource so when a contact came online there'd be a few
ClientTypesUpdated signals fired with no information because
GabblePresence thought we didn't know anything about the types, when
in actual fact we just hadn't looked yet.

This sets a resource's client types to a silly value so that it's
obvious when it hasn't been set and shows that CTUpdated shouldn't be
fired in this case.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>