tests: print out whether a core file exists in red
Previously, as it turns out, if the test(s) failed, the execution of
the check-twisted rule would go no further, so it wouldn't let us know
when a core file existed.

This patch not only fixes this behaviour so that not only will the
"core dump exists" string will get printed, but it will be in red so
the astute gabble developer will see it and profit from it.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>