Rename --with-ca-file to --with-ca-certificates
This is the option name that Gabble historically used, and is somewhat
more accurate. For GNUTLS the system CA store is a single concatenated
file ("CA bundle"). OpenSSL can use a CA bundle too, but it's more
conventional and efficient to use a directory of files/symlinks named
according to their fingerprints, which can be opened lazily; in this case
you'd use (for instance) --with-ca-certificates=/etc/ssl/certs on
Debian-derived systems.

When auto-detecting, we still only look for files, since they work with
any backend, and the auto-detection doesn't know which backend is in use.
Packagers should probably always specify a path anyway.

Conflicts:

	configure.ac