tests: Support session establishment.
XMPP IM defines a "session establishment" stream feature:
<http://xmpp.org/rfcs/rfc3921.html#session>. This is optional, and Wocky
deals correctly with it being unsupported, but prpl-jabber assumes it's
supported and just blindly sends it. We should send *a* reply to the IQ,
whether success or an error, per XMPP Core; we may as well implement it
properly.

(I'm filing a bug against prpl-jabber just as soon as I recover my
Pidgin Trac account, but still.)