Move zombie requests into the crypt
If we keep our dear living dead in the active lists we tend to count them the
same as the living, which means the pipeline stalls even if requests time out.

Also don't accidentally turn requests that are still pending into zombies,
directly delete them instead. Otherwise we accidentally send the zombies out
into the world and trigger the zombie apocolypse (Or, well, cause a gabble
crash when the zombie times out and the original callback gets called and
gnawed on)