gabbletest: silence the queue if we hit an error
Previously, if you had the queue on verbose (CHECK_TWISTED_VERBOSE=1)
and you hit an assertion or timeout, the traceback would be printed,
followed by hundreds more events that occurred as a result of the
disconnection. This makes finding the actual problem quite hard.

This patch silences the queue once an exception is thrown so the last
thing in the verbose log is the traceback.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>