Search: handle Server='' as if it were omitted entirely
Previously, if you included the Server property in a channel request,
but set it to the empty string, Gabble would try to send the search to
the empty JID.

Instead, it ought to either reject the request outright, or pretend that
the property was omitted. I chose the latter: it's kinder to application
authors.

Fixes: <https://bugs.freedesktop.org/show_bug.cgi?id=32390>