Location: don't poll for people's locations.
As discussed on <https://bugs.freedesktop.org/show_bug.cgi?id=27693>,
the point of PEP is that you never have to poll for people's latest
location if you advertise the relevant +notify capability. So we
shouldn't do this in response to calls to GetLocations() or
GetContactAttributes().

Fixing <https://bugs.freedesktop.org/show_bug.cgi?id=31218> will
essentially re-add this code...