roster: test handling pushes without id='' attribute.
I previously believed that I had added code to WockyPorter such that, if
someone sent us an IQ with no id='' attribute, it would be silently
dropped rather than being handed up to the application. (This would
technically be valid behaviour; RFC3920 §9.2.3 “IQ Semantics” says that
“The 'id' attribute is REQUIRED for IQ stanzas.”
<http://xmpp.org/rfcs/rfc3920.html#stanzas-semantics-iq>)

I mentioned this to Rob, who was horrified, because Gabble's roster code
specifically handles roster pushes lacking an id='' attribute because
this happens in the wild. I checked, and I had apparently imagined that
I had made the above change to WockyPorter.

This patch adds a regression test to check that Gabble correctly
processes roster pushes without an id='' attribute, and doesn't crash
in the process.