Roster: respond to malicious pushes we drop.
Gabble correctly does not process roster pushes from contacts other than
our server. However, it does not send a reply back to the IQ, which is
in violation of XMPP Core.

This patch allows the IQ to fall through to the “no handler” code, which
just sends back a generic IQ error. While we could craft a specific
<forbidden/> error, I don't think it's necessary.