contact-info: don't clobber the per-process array of vcard fields depending on server type
Previously, when setting the Overwritten_By_Nickname flag on the
FN/NICKNAME VCard field it was done in the per-process array.

As a result, if you signed on with a Gtalk account, FN will be fiven
said flag. If you then sign on with a non-Google account, the FN field
will already have this flag set so both FN and NICKNAME will have it
set, which is wrong.

This is never a problem the other way round because Gtalk doesn't
support the NICKNAME VCard field.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>