Add support for ServerTLSConnection.ReferenceIdentities property.
The certificate identity can be checked against more than just
one piece of information. In gabble we have two such possible
identities. The JID domainpart, and (if specified) the XMPP server
overridden in configuration. Both of these are explicitly configured
by a user and therefore can be used in certificate identity checks.

https://bugs.freedesktop.org/show_bug.cgi?id=35410