MUC: deal with failing to join, or being kicked.
ffc594c, which made leaving rooms based on user request asynchronous,
unfortunately meant that the case where the server refuses our entry to
the room in the first place, or unilaterally removes us from a room
we're in, crashed Gabble.

This was because the call to tp_base_channel_destroyed() calls out to
various bits of code which ultimately leaves the MucChannel with no
references, so it gets disposed … right before we start calling methods
on it. This wasn't a problem previosuly because
tp_base_channel_destroyed() was the last function called in this
function.

This patch is a bit of a band-aid: it just refs the channel across the
duration of this function. It's the least invasive solution that works.
I've added a couple of tests: one for failing to join a room (which was
filed as fd.o#35120), and another for being kicked from a room (which
was previously untested, and triggered this same bug).