tests: include test directory in resource
gabbletest.py sets the resource in the 'account' parameter to the name
of the current test case; it's overridden by the explicit 'resource'
parameter, so has no effect, but shows up in gabble-testing.log, making
it easier to find the section of the log pertaining to the failing test
case.

Commit 903ed6c made the resource 'nicer' by removing everything except
the last component of the test name, so rather than using (for instance)
'text/send-error.py' as the resource, 'send-error.py' was used. This
doesn't seem 'nicer' to me: there are two tests called 'send-error.py',
one in 'text/' and one in 'muc/'.

This commit also strips leading './', if present: my `tt` test runner
ends up putting these in (as part of its handling for out-of-tree
builds).

Reviewed-by: Jonny Lamb <jonny.lamb@collabora.co.uk>