Clarify resource aggregration's preferences
When I looked at the changes to aggregate_resources() introduced by
8945d56, I thought that a regression (for which I've added a test) might
have occurred because the 'prio' and 'activity' variables would not be
updated for the first resource in the list. As it happens, they are, but
only because the second 'if' was not an 'else if'.

So I've condensed them into one, and removed the extraneous variables.