test plugin: make RCC conform to the spec.
The spec requires that the fixed properties of any requestable channel
class include ChannelType and TargetHandleType. (Actually, we violate
the latter for ContactSearch, but the former seems like a pretty
reasonable thing to conform to.) The test plugin did not respect this:
in fact, none of the fixed or allowed properties were valid D-Bus
property names.

Anyway, some of the Jingle tests rely on ChannelType being present in
every channel class. This seems basically reasonable to me.

I did not catch this in my review.