Presence: rename Resource.last_activity to last_available
This field is not actually the time of the last activity of any kind by
that resource; it's not even the time of the last presence received for
that resource. It's the time of the last *available* presence received
for that resource! It's been this way since the dark days of darcs
(417c27a) and no-one's complained, so I assume that it's intentional;
but it seems like we should rename the field.