Presence: turn client types into a bitfield sooner.
There's no point in turning these into a GPtrArray of dup'd strings if
the only thing we do with that array is to turn it into a bitfield.
Turning them straight into the bitfield makes the memory management
simpler (i.e. there is none).

(Later, it gets turned back into a GPtrArray! Astonishing.)