hashed caps tests: correctly handle identity languages
The helper function to build disco replies ignored the language field of
the identity strings passed to it. This meant that any tests which
happened to use identities with non-empty languages should fail, because
Gabble would (correctly) think that the hash doesn't match the reply.

The test that actually sends hashed caps to/from Gabble didn't use any
identities at all (in violation of the MUST in XEP-0030), so this wasn't
noticed.