ClientTypes: fix satisfying two waiters with one reply
Previously, if two contacts had the same previously-unknown hash, we'd
only disco one of them (correctly), but when the result came back we
would only update the client types for the contact who replied, not the
second.

This patch adds a regression test, and fixes this issue by rolling the
client type propagation into some of the existing code to update caps.
I think it should really go further—the serial stuff applies equally to
identities as to features—but life is too short, and clients should
(presumably) return the same identities from all disco queries.