Clarify get_client_types_from_handle()
I actually broke the code path that tries to avoid signalling the empty
list when a disco request is in progress:
gabble_presence_get_client_types_array() always returns non-NULL these
days. This path wasn't exercised, but it's about to be… It took me quite
a few readings to figure out exactly what the function was trying to do,
so this patch clarifies it.