gabbletest: disconnect catch-all-signals match rule
Previously, the catch-all-signals match rule (and callback) was never
disconnected. This doesn't matter for tests in separate processes, but
if you have n tests in sequence in one file, you'd end up with n²/2
copies of the callback. This slows down the test, and (more importantly)
prevents most of the test harness from being freed: events get added to
the queues for previous tests.

I only noticed this because I wrote a test case that runs in a loop in a
single process. The python process ballooned to over 2 gigabytes, and
then died with ‘dbus.exceptions.DBusException:
org.freedesktop.DBus.Error.LimitsExceeded: Connection ":1.0" is not allowed to
add more match rules (increase limits in configuration file if
required)’ when trying to add just one more wafer-thin match rule.