servicetest: disable unnecessary introspection
dbus-python's automatic introspection is actually pretty costly.
Profiling connect/torture.py shows that it's responsible for about 10%
of the runtime of that test.

The majority of that is introspecting the connection, and sadly many,
many tests depend on that working. But we can at least turn it off on
the dummy object used for sync_dbus, and for the CM object used to
request the connection. This shaves a few seconds off the test suite.
Not zero, but makes no real impact on the six minutes the test suite
takes. :(