Explicitly cast to guint64 when passing varargs
Without any additional type information, integers used in varargs lists
can get passed as the wrong type. In this call to tp_value_array_build()
in gabble_muc_channel_constructed(), we should be passing 0 as a a
gint64, but I think in practice we're passing it as a pointer-sized
thing. (I'm not a C language lawyer, I'm sure someone could correct me
on this.) If the type we're passing it as is not the same size as a
gint64, this will cause tp_value_array_build() to read off the end of
its varargs, probably crashing in the process (because what it reads
will probably not be a valid GType).

The other call to tp_value_array_build() added in ec8ca6e (for the Room
interface) does cast correctly.

Fixes: <https://bugs.freedesktop.org/show_bug.cgi?id=37165>
Reviewed-by: Jonny Lamb <jonny.lamb@collabora.co.uk>