Update Wocky snapshot for WockyMuc changes
This update changes the signature of various signals to include MUC
status codes as bits in a guint rather than in a hash table. It also
fixes WockyMuc stealing <presence/> stanzas before Gabble's handlers
have a chance to get stuff out of them; test for that to follow.

Fixes: <https://bugs.freedesktop.org/show_bug.cgi?id=32017> but I am not
very happy with the fix.