Look for better aliases if roster says name=jid.
When we discover that a contact doesn't have an alias at all, we fall
back to their JID, and write this to the roster (as we do for any other
alias we use, to spare our users' precious bandwidth).

This is fine and all, but it means that we ignore any subsequent updates
from the contact to say that they have a better alias. For instance, we
might get a PEP nickname update, or fetch their vCard for some other
reason (getting an avatar, ContactInfo, &c &c). In those cases, we would
like to have any better alias obtained passively to replace the JID
fallback. This is only ever passive: we still don't query for a better
alias if the roster says it's the JID. (Clever users can set the alias
to the empty string, which has the effect of forcing a refresh.)

Fixes: https://bugs.freedesktop.org/show_bug.cgi?id=27361