test-save-alias-to-roster: ack and echo roster pushes
I was adding some more stuff to this test and realised that it's
technically misbehaving as a server: the server MUST (of course) ack the
roster update the client says, but it also MUST send roster pushes to
all connected resources, event to the resource who made the change.
Gabble relies on this, and only respects changes to the roster when the
server pushes them to it. (I think this is fine, it means there's only
one code path.)

This test happened not to ever update a roster item more than once, so
not acking the pushes didn't matter so much. It also never depended on
reading back stuff that it added to the roster. But I'm about to add
a section which will.