ContactInfo: don't claim PHOTO is unknown
Previously this code would log “unknown vCard node in XML: PHOTO” for
every vCard that flew past. This is misleading: PHOTO is perfectly
well-known, it's just handled by a different bit of code. So this patch
adds the notion of an ignored field, which the ContactInfo code knows
about but does nothing with.

I didn't add any of the other fields mentioned in the comment at the end
of the field table: they're not handled anywhere, so it may arguably be
worth logging something about these.