PresenceCache: discard UNKNOWN presences in maybe_remove
In various places, gabble_presence_cache_maybe_remove() is called to
allow the cache to relinquish its grip on content-free offline
presences. It should also be able to drop unknown presences: they're
basically equivalent, based on whether or not the contact is on our
roster.

Specifically, this means that UNKNOWN presences for contacts received
before the roster has arrived do not linger in the cache, and hence
these contacts are correctly signalled as OFFLINE when the roster
arrives.

Really, I think a better fix would be to change the presence cache to
not have this distinction between Unknown and Offline: it's completely
based on whether or not the contact is (known to be) on the roster, and
this is reflected in the way that the handler for incoming <presence/>
passes a different value into the presence cache based on whether or not
the contact is on the roster. But this would be a much more invasive
change, so I'm not prepared to make it just yet.