Aliasing: split out cached remote alias lookup
There are no code changes in this patch: it simply moves the middle
section of _gabble_connection_get_cached_alias(), which deals with
looking for cached aliases that the contact specified for themself (on
PEP, in their vCard, in their MUC JID, etc.) into a separate function.

It also adds a docstring to _gabble_connection_get_cached_alias().