Aliasing: handle SetAlias({contact: ''}) better
Previously, if the user set a contact's alias to '', Gabble would change
its alias on your roster to the empty string… this is not really ideal.

This patch makes Gabble switch to using a cached remote alias for the
contact, if any; if there is none, it simply removes the name=''
attribute from the roster entirely, and tries to make a request for a
better nickname.