Add an extra_identify argument to the TLS verification functions
Update the Wocky snaphost and adapt the TLS manager to the changes.

Now the TLS verification functions have an extra argument for the
additional hostnames that we have to consider valid.
This is useful, for instance, when using Google apps. Rather than
providing a certificate for the JID’s domain part, the server provides a
certificate for talk.google.com; if the user has explicitly configured a
‘Google Talk’ account, it's reasonable to accept certificates for this
domain.