Test Debug object's absence with --disable-debug
38a6178 changed test-debug.py to verify that the Debug object is present
but non-functional when --disable-debug is passed to configure. Gabble
actually doesn't have a Debug object at all in that situation, which I
think is reasonable—so this now tests that calling GetMessages fails.
The test doesn't make any particular assertions about the error. Right
now it happens to be 'org.freedesktop.DBus.Error.UnknownMethod' but it
really ought to be some kind of ‘unknown object’ method. So.