MUC: don't forget password when handling nick conflicts
WockyMuc has a property, :password, representing the current password
being used to join the MUC. GabbleMuc previously had a private variable,
'password', which was used for this. In the port to WockyMuc, setting
the private variable was removed, but it was still used when
re-attempting to join after a nick conflict.

(I think the password should be a parameter to a hypothetical
wocky_muc_join_async() which does all the nick conflict crap for you.
Having this as state that kicks around on the WockyMuc for ever is
bizarre—once you're in the room, you usually don't use the password,
unless you're the owner, in which case you can retrieve the current
password *which may be different*!)

This patch expunges the zombie private variable, and ensures
WockyMuc:password is only set when the user provides a password, not at
every join attempt. It adds a test for this case, and some of the basic
functionality of Password (which subsumes some incidental testing of the
Password interface in muc/presence-before-closing: the only test that
touched Password at all!).

Fixes: https://bugs.freedesktop.org/show_bug.cgi?id=39790
Reviewed-by: Jonny Lamb <jonny.lamb@collabora.co.uk>