Connection: simplify getting ContactCaps without presence
If we don't have presence for a contact, we do not have any capabilities
for them. But we always want to ensure that all contacts are capable of
IM. So, currently there is a special-case in the caps-gathering code: if
we have no presence, we only call _get_contact_capabilities on the IM
factory: its implementation of this virtual method, unlike all others,
doesn't care if it's passed a NULL set of caps. Otherwise, if we do have
a set of capabilities, we call gabble_connection_build_contact_caps(),
which is essentially a foreach over all channel managers, including the
IM factory.

I think it's neater to avoid the special case, and treat "no presence"
as equivalent (for capabilities purposes) to "presence with no caps".