MUC: remove half-baked allowinvites support
Previously, we allowed the user to configure whether or not occupants
should be allowed to invite others to the room, but didn't actually
provide any way to retrieve the current setting.

This is a pretty useless MUC configuration flag anyway, given that
occupants can always send direct (as opposed to mediated)
invitations … so I have no qualms about expunging it.