MUC: update for RoomConfig.update_async API change
I decided it was clearer for TpBaseRoomConfigClass.update_async to take
a TpBaseRoomConfig as its first argument, rather than a TpBaseChannel.

This makes it reasonable to implement the update_async vfunc in
GabbleRoomConfig! Yay, no more monkey-patching.

I think it would be even better to move all the configuration-frobbing
code to GabbleRoomConfig. But this branch is already plenty long enough.