MUC: use id='' to catch SetSubject errors
Servers don't strictly have to echo the <subject/> element in their
reply if they reject our subject change message. But they do have to
preserve the id='' attribute.

This patch makes Gabble set an id='' attribute on <message/>s sent to
change the subject, and look for that ID when it receives a <message
type='error'/> from a MUC. (It continues to look for <subject/> errors,
too.)

Prosody is an example of a server that behaves like this.