Call: test google relay reply after channel dies
It's kind of funny: the bug fixed by the previous commit repeatedly
surfaced for StreamedMedia channels in a variety of situations (you
closed the channel, you removed yourself from Members, you disconnected
the connection), and hence there's a lovingly-written exhaustive set of
test cases for those scenarios for incoming and outgoing channels. But
of course they weren't run for Call channels…

This patch only makes the bare minimum number of changes to the test to
at least slightly exercise the Call code. It's not as exhaustive as it
is for StreamedMedia, and works almost by accident, but hey...