Fix offline contacts showing up as unknown, not offline
Due to a weird interaction between the presence cache, IM channels, and
scraping nicknames out of <message/>s, receiving a message from an
offline contact before the roster is received would cause their status
to erroneously show up as unknown, not offline.

This fix is a bit of a hack, but it is much smaller than refactoring to
make the IM channel store the alias (which would allow us to expunge
keep_unavailable).

This regressed as a side-effect of e0cda61.

Fixes: <https://bugs.freedesktop.org/show_bug.cgi?id=41743>