Test grabbing <nick/> from <message/>
If a contact is not on your roster, you typically have no idea what
their nickname is: no roster, no PEP, no vCard (assuming the server
doesn't let random people fetch your vCard). In this situation, contacts
who message you out of the blue can include <nick/> in the message
itself.

This is implemented in a kind of dodgy way in Gabble at the moment: the
IM channel forcibly retains an entry for the contact in the presence
cache, and then the presence cache stashes the nickname as if it came
from presence… It was also previously untested, so I thought it worth
adding a test before I even thought about fixing how it's implemented.