roster: ignore multiple replies to roster query.
The XMPP server running on vk.com is buggy, and replies to our roster
query twice. Gabble just blindly assumed that any <iq type='result'>
with a roster in it was the singular reply to our singular query. This
is a pretty reasonable assumption but this buggy server violates it, and
Gabble calls tp_base_contact_list_set_list_received() more than once,
which triggers a critical in that function.

A more invasive fix would use wocky_porter_send_iq_async() rather than
scraping all the roster IQs out of everything using the LM API, but I
wanted to quickly fix a bug on a Friday afternoon.

Fixes: <https://bugs.freedesktop.org/show_bug.cgi?id=42186>
Reviewed-by: Jonny Lamb <jonny.lamb@collabora.co.uk>