im_channel_receive: remove redundant 'sender' parameter
If we're passing an incoming message to a 1-1 IM channel, we don't need
to tell the channel who sent the message: it knows perfectly well. In
fact, we were telling it twice: once in handle form (which is
redundant), and once in full JID form (which is more useful: it lets the
channel update its resource binding).