ImFactory: handle Facebook's own-message extension
Facebook's XMPP server sends us IQs containing messages which we send to
contacts, whether on this connection or via another connection (such as
the web interface). We can expose these as Delivery_Status_Accepted
delivery reports, allowing UIs to log them and show them up or whatever.
This could help keep things coherent if the user hops between chatting
in the browser and/or Empathy, say.

https://bugs.freedesktop.org/show_bug.cgi?id=41417