conn-olpc: handle activities with NULL ids.
Other places in the code seem to handle GabbleOlpcActivity.id being
NULL; and <https://bugs.freedesktop.org/show_bug.cgi?id=32050> is a
crash on this strcmp. There are no logs, and the first argument to
strcmp is not on record, so this suggests to me that this might be the
cause!