Presence: show better error status messages.
When you get <presence type='error'>, any <status/> element in the
presence is an echo of a stanza you sent out. So the message in that is
not the contact's presence: it's either your status, or maybe the
message you sent with a subscription request, depending on whether we
got the <presence type='error'/> for a roster contact or someone we
tried to subscribe to.

So instead, I think we should use the text of the error (if any) as the
status for such contacts, falling back to the error element name if the
server is unkind and doesn't include an error message. Yes, this means
strings will show up in the UI in the server's locale, but this is
hardly news.

https://bugs.freedesktop.org/show_bug.cgi?id=45491
Reviewed-by: Simon McVittie <simon.mcvittie@collabora.co.uk>